@@ -8,39 +8,39 @@
         "ext-pdo_mysql": "*",
         "doctrine/cache": "^1.10",
         "doctrine/doctrine-bundle": "^2.3",
-        "doctrine/orm": "^2.8",
+        "doctrine/orm": "^2.9",
         "erusev/parsedown-extra": "^0.8.1",
         "knplabs/doctrine-behaviors": "^2.2",
         "nyholm/psr7": "^1.4",
         "rector/rector": "^0.11.1",
         "sensio/framework-extra-bundle": "^5.6",
         "sentry/sentry-symfony": "^4.0",
-        "symfony/asset": "^5.2",
-        "symfony/cache": "^5.2",
-        "symfony/console": "^5.2",
-        "symfony/debug-bundle": "^5.2",
-        "symfony/dependency-injection": "^5.2",
-        "symfony/dotenv": "^5.2",
-        "symfony/event-dispatcher": "^5.2",
-        "symfony/form": "^5.2",
-        "symfony/framework-bundle": "^5.2",
-        "symfony/http-kernel": "^5.2",
-        "symfony/intl": "^5.2",
-        "symfony/mailer": "^5.2",
-        "symfony/mailgun-mailer": "^5.2",
-        "symfony/mime": "^5.2",
-        "symfony/process": "^5.2",
-        "symfony/security-bundle": "^5.2",
-        "symfony/security-core": "^5.2",
-        "symfony/templating": "^5.2",
-        "symfony/translation": "^5.2",
-        "symfony/twig-bridge": "^5.2",
-        "symfony/twig-bundle": "^5.2",
-        "symfony/uid": "^5.2",
-        "symfony/validator": "^5.2",
-        "symfony/web-profiler-bundle": "^5.2",
+        "symfony/asset": "^5.3",
+        "symfony/cache": "^5.3",
+        "symfony/console": "^5.3",
+        "symfony/debug-bundle": "^5.3",
+        "symfony/dependency-injection": "^5.3",
+        "symfony/dotenv": "^5.3",
+        "symfony/event-dispatcher": "^5.3",
+        "symfony/form": "^5.3",
+        "symfony/framework-bundle": "^5.3",
+        "symfony/http-kernel": "^5.3",
+        "symfony/intl": "^5.3",
+        "symfony/mailer": "^5.3",
+        "symfony/mailgun-mailer": "^5.3",
+        "symfony/mime": "^5.3",
+        "symfony/process": "^5.3",
+        "symfony/security-bundle": "^5.3",
+        "symfony/security-core": "^5.3",
+        "symfony/templating": "^5.3",
+        "symfony/translation": "^5.3",
+        "symfony/twig-bridge": "^5.3",
+        "symfony/twig-bundle": "^5.3",
+        "symfony/uid": "^5.3",
+        "symfony/validator": "^5.3",
+        "symfony/web-profiler-bundle": "^5.3",
         "symfony/webpack-encore-bundle": "^1.8",
-        "symfony/yaml": "^5.2",
+        "symfony/yaml": "^5.3",
         "symplify/amnesia": "^9.3",
         "symplify/autowire-array-parameter": "^9.3",
         "symplify/composer-json-manipulator": "^9.3",
@@ -60,8 +60,8 @@
         "phpstan/phpstan-doctrine": "^0.12.33",
         "phpunit/phpunit": "^9.5",
         "roave/security-advisories": "dev-master",
-        "symfony/browser-kit": "^5.2",
-        "symfony/css-selector": "^5.2",
+        "symfony/browser-kit": "^5.3",
+        "symfony/css-selector": "^5.3",
         "symplify/coding-standard": "^9.3",
         "symplify/easy-coding-standard": "^9.3",
         "symplify/phpstan-extensions": "^9.3",
